AI-Agents/AI-Team/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.45.0
google-genai==1.56.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-google-genai==4.1.2
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
marshmallow==3.26.2
mcp==1.26.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
openai==2.14.0
openrouter==0.1.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
pypdf==6.5.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.45
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.41.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
